+++ changelog.md
@@ -1,5 +1,5 @@
 ### General
 
 - Added warning for `frostWalkerNoFreezing` as enchantments are defined by datapacks in 1.21+
-- Made address weight reduce over time
+- Address weight now reduces over time
 - Performance and experience improvements